+++ server/src/main/java/dataaccess/DataAccess.java
@@ -1,4 +1,8 @@
 package dataaccess;
 
+import dataaccess.dao.IntAuthDAO;
+import dataaccess.dao.IntGameDAO;
+import dataaccess.dao.IntUserDAO;
+
 public record DataAccess(IntAuthDAO authDAO, IntGameDAO gameDAO, IntUserDAO userDAO) {
 }
