+++ server/src/main/java/dataaccess/SqlUserDAO.java
@@ -1,5 +1,6 @@
 package dataaccess;
 
+import dataaccess.dao.IntUserDAO;
 import model.UserData;
 import org.mindrot.jbcrypt.BCrypt;
