+++ server/src/main/java/dataaccess/dao/IntGameDAO.java
@@ -1,5 +1,6 @@
-package dataaccess;
+package dataaccess.dao;
 
+import dataaccess.DataAccessException;
 import model.GameData;
 
 import java.util.Collection;
