+++ server/src/main/java/dataaccess/dao/IntUserDAO.java
@@ -1,5 +1,6 @@
-package dataaccess;
+package dataaccess.dao;
 
+import dataaccess.DataAccessException;
 import model.UserData;
 
 public interface IntUserDAO {
