+++ server/src/main/java/dataaccess/dao/MemAuthDAO.java
@@ -1,4 +1,4 @@
-package dataaccess;
+package dataaccess.dao;
 
 import model.AuthData;
