+++ server/src/main/java/dataaccess/dao/MemUserDAO.java
@@ -1,4 +1,4 @@
-package dataaccess;
+package dataaccess.dao;
 
 import model.UserData;
