+++ server/src/test/java/service/UserServiceTest.java
@@ -1,6 +1,7 @@
 package service;
 
 import dataaccess.*;
+import dataaccess.dao.*;
 import model.AuthData;
 import model.GameData;
 import model.UserData;
